Gewicht loggen

+++ Workout/HttpClients/GoogleClient.cs
@@ -58,7 +58,7 @@ public class GoogleClient(HttpClient client, IOptions<JsonSerializerOptions> jso
         }
     }
 
-    public Task SavePersoonlijkeGegevensLogAsync(int gewicht)
+    public Task SavePersoonlijkeGegevensLogAsync(decimal gewicht)
     {
         var fileName = $"{DateTime.Now:yyyy-MM-dd} persoonlijke info.json";
